Cast to int

+++ SU2_PY/SU2/amginria/interface.py
@@ -209,22 +209,22 @@ def read_mesh_and_sol(mesh_name, solution_name):
     
     amgio.py_ReadMeshAndSol(mesh_name, solution_name, Ver, Cor, Tri, Tet, Edg, Hex, Qua, Pyr, Pri, Sol, SolTag,  Markers)
         
-    NbrTet = len(Tet)/5
+    NbrTet = int(len(Tet)/5)
     Tet = np.reshape(Tet,(NbrTet, 5)).astype(int)
     
-    NbrTri = len(Tri)/4
+    NbrTri = int(len(Tri)/4)
     Tri = np.reshape(Tri,(NbrTri, 4)).astype(int)
     
-    NbrEdg = len(Edg)/3
+    NbrEdg = int(len(Edg)/3)
     Edg = np.reshape(Edg,(NbrEdg, 3)).astype(int)
 
-    NbrCor = len(Cor)
+    NbrCor = int(len(Cor))
     Cor = np.reshape(Cor,(NbrCor, 1)).astype(int)
 
-    NbrVer = len(Ver)/3
+    NbrVer = int(len(Ver)/3)
     Ver = np.reshape(Ver,(NbrVer, 3))
     
-    SolSiz = len(Sol)/NbrVer
+    SolSiz = int(len(Sol)/NbrVer)
     Sol = np.array(Sol).reshape(NbrVer,SolSiz).tolist()
     
     # First row of Markers contains dimension
